ISSUE #3 rename: Rename Tab components
